Fixed 15 -Wsign-compare] warnings
either through explicit casts, or changing the underlying types where needed.

I also believe I uncovered one existing bug (see line 527 of p_pspr.c), but I haven't attempted a fix or taken a closer look at what affect this possibly buggy code may have.

+++ src/d_server.c
@@ -234,7 +234,7 @@ int displaycounter;
 boolean n_players_in_state(int n, int ps) {
   int i, j;
   for (i = j = 0; i < MAXPLAYERS; i++)
-    if (playerstate[i] == ps)
+    if ((int)playerstate[i] == ps)
       j++;
   return (j == n);
 }
