+++ src/gl_main.c
@@ -386,7 +386,7 @@ void gld_Init(int width, int height) {
       while (*p && *p != ' ')
         p++;
       if (*p) {
-        int len = MIN(p - rover, sizeof(ext_name) - 1);
+        int len = MIN((size_t)(p - rover), sizeof(ext_name) - 1);
         memset(ext_name, 0, sizeof(ext_name));
         strncpy(ext_name, rover, len);
         lprintf(LO_INFO, "\t%s\n", ext_name);
